+++ transactions.go
@@ -1,0 +1,82 @@
+package pokemonbattlelib
+
+import "fmt"
+
+// Transactions describes a change to battle state.
+// A sequence of transactions should be able to describe an entire battle.
+type Transaction interface {
+	BattleLog() string
+}
+
+// A transaction to deal damage to an opponent Pokemon.
+type DamageTransaction struct {
+	User            *Pokemon
+	Target          *Pokemon
+	TargetParty     int
+	TargetPartySlot int
+	Move            *Move
+	Damage          uint
+}
+
+func (t DamageTransaction) BattleLog() string {
+	return fmt.Sprintf("%s used %s on %s for %d damage.",
+		t.User.GetName(),
+		t.Move.Name,
+		t.Target.GetName(),
+		t.Damage,
+	)
+}
+
+// A transaction to use and possibly consume an item.
+type ItemTransaction struct {
+	Target *Pokemon
+	Item   *Item
+	Move   *Move
+}
+
+func (t ItemTransaction) BattleLog() string {
+	return fmt.Sprintf("%s used on %s.", t.Item.Name, t.Target.GetName())
+}
+
+// A transaction to restore HP to a Pokemon.
+type HealTransaction struct {
+	Target *Pokemon
+	Amount uint
+}
+
+func (t HealTransaction) BattleLog() string {
+	return fmt.Sprintf("%s restored %d HP.", t.Target.GetName(), t.Amount)
+}
+
+// A transaction that makes a pokemon faint, and returns the pokemon to the pokeball.
+type FaintTransaction struct {
+	Target          *Pokemon
+	TargetParty     int
+	TargetPartySlot int
+}
+
+func (t FaintTransaction) BattleLog() string {
+	return fmt.Sprintf("%s fainted.",
+		t.Target.GetName(),
+	)
+}
+
+// A transaction that makes a party send out a pokemon.
+type SendOutTransaction struct {
+	Target          *Pokemon
+	TargetParty     int
+	TargetPartySlot int
+}
+
+func (t SendOutTransaction) BattleLog() string {
+	return fmt.Sprintf("%s was sent out.",
+		t.Target.GetName(),
+	)
+}
+
+type EndBattleTransaction struct{}
+
+func (t EndBattleTransaction) BattleLog() string {
+	// TODO: include reason the battle ended
+	return "The battle has ended."
+}
